Refactor: user, inputvalue 상태 전역으로 관리
